Move embedded generation of the interceptable attribute to the interceptor package
