+++ src/NetEscapades.EnumGenerators.Interceptors/SourceGenerationHelper.cs
@@ -15,6 +15,49 @@ public static class SourceGenerationHelper
 
 #nullable enable";
 
+    public const string Attribute =
+        $$"""
+          {{Header}}
+
+          #if NETESCAPADES_ENUMGENERATORS_EMBED_ATTRIBUTES
+          namespace NetEscapades.EnumGenerators
+          {
+              /// <summary>
+              /// Add to an assembly to indicate that usages of the enum should
+              /// be automatically intercepted to use the extension methods
+              /// generated by EnumExtensionsAttribute in this project.
+              /// Note that the extension methods must be accessible from this project,
+              /// otherwise you will receive compilation errors
+              /// </summary>
+              [global::System.AttributeUsage(global::System.AttributeTargets.Assembly, AllowMultiple = true)]
+              [global::System.Diagnostics.Conditional("NETESCAPADES_ENUMGENERATORS_USAGES")]
+          #if NET5_0_OR_GREATER
+              [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage(Justification = "Generated by the NetEscapades.EnumGenerators.Interceptors source generator.")]
+          #else
+              [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+          #endif
+              public class InterceptableAttribute<T> : global::System.Attribute
+                  where T: global::System.Enum
+              {
+                  /// <summary>
+                  /// The namespace generated for the extension class. If not provided,
+                  /// and the referenced enum is in a different project, the namespace
+                  /// of the extension methods are assumed to be the same as the enum.
+                  /// </summary>
+                  public string? ExtensionClassNamespace { get; set; }
+                  
+                  /// <summary>
+                  /// The name used for the extension class. If not provided,
+                  /// and the referenced enum is in a different project, the enum name
+                  /// with an <c>Extensions</c> suffix will be assumed. For example for
+                  /// an Enum called StatusCodes, the assumed name will be StatusCodesExtensions.
+                  /// </summary>
+                  public string? ExtensionClassName { get; set; }
+              }
+          }
+          #endif
+          """;
+
     public static (string Content, string Filename) GenerateInterceptorsClass(MethodToIntercept toIntercept)
     {
         var sb = new StringBuilder(
