+++ src/NetEscapades.EnumGenerators/SourceGenerationHelper.cs
@@ -92,38 +92,6 @@ namespace NetEscapades.EnumGenerators
         /// </summary>
         public bool IsInterceptable { get; set; } = true;
     }
-
-    /// <summary>
-    /// Add to an assembly to indicate that usages of the enum should
-    /// be automatically intercepted to use the extension methods
-    /// generated by <see cref=""EnumExtensionsAttribute""/> in this project.
-    /// Note that the extension methods must be accessible from this project,
-    /// otherwise you will receive compilation errors
-    /// </summary>
-    [global::System.AttributeUsage(global::System.AttributeTargets.Assembly, AllowMultiple = true)]
-    [global::System.Diagnostics.Conditional(""NETESCAPADES_ENUMGENERATORS_USAGES"")]
-#if NET5_0_OR_GREATER
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage(Justification = ""Generated by the NetEscapades.EnumGenerators source generator."")]
-#else
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-#endif
-    public class InterceptableAttribute<T> : global::System.Attribute
-        where T: global::System.Enum
-    {
-        /// <summary>
-        /// The namespace generated for the extension class.
-        /// If not provided, the namespace of the enum will be assumed.
-        /// </summary>
-        public string? ExtensionClassNamespace { get; set; }
-
-        /// <summary>
-        /// The name used for the extension class.
-        /// If not provided, the enum name with """"Extensions"""" will be assumed.
-        /// For example for an Enum called StatusCodes, the assumed name
-        /// will be StatusCodesExtensions.
-        /// </summary>
-        public string? ExtensionClassName { get; set; }
-    }
 }
 #endif
 ";
